The size of bakebot was exploding to the point that it needed to be separated into its own component.  Everything works and appears as it did before, but now it's properly compartmentalized.
